Move EmptyElseBlockInspectionTests for QuickFixes into relevant files under RD.QuickFixes
