+++ RubberduckTests/Inspections/EmptyElseBlockInspectionTests.cs
@@ -4,7 +4,6 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Rubberduck.VBEditor.SafeComWrappers.Abstract;
 using RubberduckTests.Mocks;
 using Rubberduck.Inspections.Concrete;
-using Rubberduck.Inspections.QuickFixes;
 using Rubberduck.Parsing.Inspections.Resources;
 
 namespace RubberduckTests.Inspections
@@ -243,38 +242,6 @@ End Sub";
             const int expectedCount = 0;
 
             Assert.AreEqual(expectedCount, actualResults.Count());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void EmptyElseBlock_QuickFixRemovesElse()
-        {
-            const string inputCode =
-@"Sub Foo()
-    If True Then
-    Else
-    End If
-End Sub";
-
-            const string expectedCode =
-@"Sub Foo()
-    If True Then
-    End If
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new EmptyElseBlockInspection(state);
-            var inspector = InspectionsHelper.GetInspector(inspection);
-            var actualResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
-
-            new RemoveEmptyElseBlockQuickFix(state).Fix(actualResults.First());
-
-            string actualRewrite = state.GetRewriter(component).GetText();
-
-            Assert.AreEqual(expectedCode, actualRewrite);
         }
     }
 }
