+++ RubberduckTests/RubberduckTests.csproj
@@ -121,6 +121,7 @@
     <Compile Include="PostProcessing\ModuleRewriterTests.cs" />
     <Compile Include="QuickFixes\ApplicationWorksheetFunctionQuickFixTests.cs" />
     <Compile Include="QuickFixes\IgnoreOnceQuickFixTests.cs" />
+    <Compile Include="QuickFixes\RemoveEmptyElseBlockQuickFixTests.cs" />
     <Compile Include="QuickFixes\RemoveUnusedDeclarationQuickFixTests.cs" />
     <Compile Include="SmartIndenter\VerticalSpacingTests.cs" />
     <Compile Include="SourceControl\SynchrounouslyConstructedDeclarationFinderFactory.cs" />
